Controlling and Selecting workers (#42)
* added functionality to  deselect/select units.

* select one or multiple units, textures added. Deselected added. TODO: Drag function.

* waiting for qt

* Textures select, deselect by string in stead of enum.

* selecting units in dra

* add drag handler, busy with lines.

* dragging and drawing

* world singleton

* initial control workers

* fixed issues, it works now

* fixes after merge

* merge mark's drag fix

* fixes, added possession check

* working selection and orders merged into dev

* fixes for Rboin.

* bug fix

+++ src/logic/entity/static/resource_entity.cpp
@@ -5,6 +5,7 @@
 #include <sdl/image/sdl_image_render_object.h>
 #include "resource_entity.h"
 #include "resource_manager.h"
+#include <string>
 
 ResourceEntity::ResourceEntity(const mesh *base, vec2 position, float mass, ResourceType rt)
         : StaticEntity(base,
@@ -25,7 +26,7 @@ ResourceType ResourceEntity::get_resource_type() {
 }
 
 int ResourceEntity::get_units() {
-    return _units;
+    return (int)_units;
 }
 
 float ResourceEntity::gather() {
